+++ src/masks.jl
@@ -251,7 +251,22 @@ indices for `A`.
 """
 forge_mask(A::AbstractMatrix, objs::MaskElement...; kwds...) = forge_mask(A, Mask(objs); kwds...)
 forge_mask(A::AbstractMatrix, msk::Mask; kwds...) =
-    forge_mask!(similar(A, floating_point_type(A)), msk; kwds...)
+    forge_mask!(similar(A, floating_point_type(eltype(A))), msk; kwds...)
+
+"""
+    TwoDimensional.forge_mask(T, dims, msk; kwds...)
+    TwoDimensional.forge_mask(T, dims, objs...; kwds...)
+
+yield a 2-dimensional array with element type `T`, dimensions `dims`, and with entries set
+to the transmission by the mask `msk`. The mask may also be specified by the list
+`objs...` of elementary mask objects. The coordinates of the mask are assumed to be given
+in fractional Cartesian indices for an array of size `dims`.
+
+"""
+forge_mask(::Type{T}, dims::Dims{2}, objs::MaskElement...; kwds...) where {T} =
+    forge_mask(T, dims, Mask(objs); kwds...)
+forge_mask(::Type{T}, dims::Dims{2}, msk::Mask; kwds...) where {T} =
+    forge_mask!(Array{T}(undef, dims), msk; kwds...)
 
 """
     TwoDimensional.forge_mask([T,] X, Y, msk; kwds...) -> arr